refactor(database): split tests files according to the refactoring
