+++ test/models/database/post_repository_test.dart
@@ -12,106 +12,6 @@ import "package:proxima/services/database/post_repository_service.dart";
 import "mock_post_data.dart";
 
 void main() {
-  group("Post Location Firestore testing", () {
-    test("hash overrides correctly", () {
-      const geoPoint = GeoPoint(40, 20);
-      const geoHash = "azdz";
-
-      final expectedHash = Object.hash(geoPoint, geoHash);
-
-      const location = PostLocationFirestore(
-        geoPoint: geoPoint,
-        geohash: "azdz",
-      );
-
-      final actualHash = location.hashCode;
-
-      expect(actualHash, expectedHash);
-    });
-
-    test("fromDbData throw error when missing fields", () {
-      final data = <String, dynamic>{
-        PostLocationFirestore.geoPointField: const GeoPoint(40, 20),
-      };
-
-      expect(
-        () => PostLocationFirestore.fromDbData(data),
-        throwsA(isA<FormatException>()),
-      );
-    });
-  });
-
-  group("Post Firestore Data testing", () {
-    test("hash overrides correctly", () {
-      final data = PostData(
-        ownerId: const UserIdFirestore(value: "owner_id"),
-        title: "post_tiltle",
-        description: "description",
-        publicationTime: Timestamp.fromMillisecondsSinceEpoch(4564654),
-        voteScore: 12,
-      );
-
-      final expectedHash = Object.hash(
-        data.ownerId,
-        data.title,
-        data.description,
-        data.publicationTime,
-        data.voteScore,
-      );
-
-      final actualHash = data.hashCode;
-
-      expect(actualHash, expectedHash);
-    });
-
-    test("fromDbData throw error when missing fields", () {
-      final data = <String, dynamic>{
-        PostData.ownerIdField: "owner_id",
-        PostData.titleField: "post_tiltle",
-        PostData.descriptionField: "description",
-        PostData.publicationTimeField:
-            Timestamp.fromMillisecondsSinceEpoch(4564654),
-      };
-
-      expect(
-        () => PostData.fromDbData(data),
-        throwsA(isA<FormatException>()),
-      );
-    });
-  });
-
-  group("Post Firestore testing", () {
-    test("hash overrides correctly", () {
-      const geoPoint = GeoPoint(40, 20);
-      const geoHash = "azdz";
-
-      const location = PostLocationFirestore(
-        geoPoint: geoPoint,
-        geohash: geoHash,
-      );
-
-      final data = PostData(
-        ownerId: const UserIdFirestore(value: "owner_id"),
-        title: "post_tiltle",
-        description: "description",
-        publicationTime: Timestamp.fromMillisecondsSinceEpoch(4564654),
-        voteScore: 12,
-      );
-
-      final post = PostFirestore(
-        id: const PostIdFirestore(value: "post_id"),
-        location: location,
-        data: data,
-      );
-
-      final expectedHash = Object.hash(post.id, post.location, post.data);
-
-      final actualHash = post.hashCode;
-
-      expect(actualHash, expectedHash);
-    });
-  });
-
   group("Post Repository testing", () {
     late FakeFirebaseFirestore firestore;
     late PostRepositoryService postRepository;
